requirementes.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
mistralai==0.4.2
websockets==12.0
pydantic==2.5.0
python-dotenv==1.0.0